requirements.txt
bert-score
faiss-gpu
nltk
pytorch
numpy
scipy